Design

+++ src/design/Employee.java
@@ -1,6 +1,6 @@
 package design;
 
-public interface Employee {
+public interface Employee{
 	
 	/*Employee is an Interface which contains multiple unimplemented methods.Again few methods
 	has been declared in below. you need to brainstorm to add more methods to meet the business requirements.
@@ -18,7 +18,7 @@ public interface Employee {
     public void assignDepartment();
 	
 	//calculate employee salary
-	public int calculateSalary();
+	public double calculateSalary();
 	
 	//employee benefit
 	public void benefitLayout();
